requirements.txt
Django==1.7
argparse==1.2.1
django-bootstrap3==5.3.1
django-debug-toolbar==1.3.0
django-guardian==1.2.5
django-reversion==1.8.6
psycopg2==2.6
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
